allows global script cache
