+++ packages/cache/package.json
@@ -21,7 +21,8 @@
     "backfill-config": "^6.3.0",
     "backfill-cache": "^5.6.1",
     "backfill-logger": "^5.1.3",
-    "fast-glob": "^3.2.11"
+    "fast-glob": "^3.2.11",
+    "glob-hasher": "1.1.1"
   },
   "devDependencies": {
     "@lage-run/monorepo-fixture": "*",
