+++ packages/cache/src/salt.ts
@@ -1,7 +1,6 @@
-import * as path from "path";
 import * as crypto from "crypto";
-import * as fg from "fast-glob";
-import * as fs from "fs/promises";
+import { hashGlobGit } from "glob-hasher";
+import { hashStrings } from "./hashStrings";
 
 interface MemoizedEnvHashes {
   [key: string]: string[];
@@ -25,6 +24,15 @@ function envHashKey(environmentGlobFiles: string[]) {
   return environmentGlobFiles.sort().join("|");
 }
 
+function sortObject(unordered: Record<string, unknown>) {
+  return Object.keys(unordered)
+    .sort((a, b) => a.localeCompare(b))
+    .reduce((obj, key) => {
+      obj[key] = unordered[key];
+      return obj;
+    }, {});
+}
+
 async function getEnvHash(environmentGlobFiles: string[], repoRoot: string) {
   const key = envHashKey(environmentGlobFiles);
 
@@ -44,37 +52,12 @@ async function getEnvHash(environmentGlobFiles: string[], repoRoot: string) {
 
 async function getEnvHashOneAtTime(environmentGlobFiles: string[], repoRoot: string) {
   const envHash: string[] = [];
-  const newline = /\r\n|\r|\n/g;
-  const LF = "\n";
-  const files = fg.sync(environmentGlobFiles, {
-    cwd: repoRoot,
-  });
 
-  files.sort((a, b) => a.localeCompare(b));
-
-  for (const file of files) {
-    const hasher = crypto.createHash("sha1");
-    hasher.update(file);
-
-    const fileBuffer = await fs.readFile(path.join(repoRoot, file), "utf-8");
-    const data = fileBuffer.replace(newline, LF);
-    hasher.update(data);
-
-    envHash.push(hasher.digest("hex"));
-  }
-
+  const hashes = hashGlobGit(environmentGlobFiles, { cwd: repoRoot, gitignore: false })!;
+  const sortedHashes = sortObject(hashes);
   const key = envHashKey(environmentGlobFiles);
-  envHashes[key] = envHash;
+
+  envHashes[key] = Object.values(sortedHashes);
 
   return envHash;
-}
-
-function hashStrings(strings: string | string[]): string {
-  const hasher = crypto.createHash("sha1");
-  const anArray = typeof strings === "string" ? [strings] : strings;
-  const elements = [...anArray];
-  elements.sort((a, b) => a.localeCompare(b));
-  elements.forEach((element) => hasher.update(element));
-
-  return hasher.digest("hex");
 }
