+++ packages/hasher/package.json
@@ -17,7 +17,8 @@
   "dependencies": {
     "execa": "5.1.1",
     "workspace-tools": "0.30.0",
-    "find-up": "^5.0.0"
+    "find-up": "^5.0.0",
+    "glob-hasher": "1.1.1"
   },
   "devDependencies": {
     "@lage-run/monorepo-fixture": "^0.1.0",
